requirements.txt
Flask
mysql-connector-python
mysql-connector
flask-login
flask-wtf
wtf-forms
pymysql
wtforms-validators